testing/web-platform/mozilla/tests/html/semantics/scripting-1/the-script-element/async-module-circular.html

<!doctype html>
<meta charset=utf-8>
<title>Load async modules circular</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
  setup({allow_uncaught_exception: true});
  var test = async_test("Load async module")
  window.addEventListener("error", scriptError);
  function scriptError(e) {
    // An error is expected
    test.step(() => assert_true(e.message !== "FAIL"));
    test.done();
  }
  function scriptLoaded() {
    test.step(() => assert_unreached("Should not load before error"));
    test.done();
  }
  function testNoError() {
    test.step(() => assert_unreached("No event expected here"));
    test.done();
  }
</script>
<script type="module"
				src="./support/async_module_circular.js"
        onerror="testNoError()"
        onload="scriptLoaded()">
</script>
